requirements.txt
numpy
scipy
Cython
h5py
mock
hope
pyephem
six
healpy
astropy
git+https://github.com/cosmo-ethz/ivy.git
